requirements.txt
requests
Flask
python-dotenv # Para cargar variables de entorno y desde .env